fix(APISIMP-REFANNOT-IN-MEMORY-PAGING-1): widen skip to long to prevent arithmetic overflow
int skip = page * pageSize overflows when page approaches Integer.MAX_VALUE (CodeQL HIGH:
"User-controlled data in arithmetic expression"). Change to long skip = (long) page * pageSize
and propagate the long type through ReferenceKindHandler.listAnnotations SPI default,
TimeseriesReferenceKindHandler, TimeseriesAnnotationDAO, VideoStreamReferenceKindHandler,
VideoAnnotationDAO, and their test files. Remove now-redundant (long) casts in the DAOs.

+++ backend/src/main/java/de/dlr/shepard/v2/references/handlers/TimeseriesReferenceKindHandler.java
@@ -219,7 +219,7 @@ public class TimeseriesReferenceKindHandler implements ReferenceKindHandler {
   }
 
   @Override
-  public List<Map<String, Object>> listAnnotations(String refAppId, int skip, int limit) {
+  public List<Map<String, Object>> listAnnotations(String refAppId, long skip, int limit) {
     return tsAnnotationDAO.findByTimeseriesReferenceAppId(refAppId, skip, limit).stream()
       .map(TimeseriesReferenceKindHandler::annotationToMap)
       .toList();
